Refactor: stage & seat refactoring
