+++ src/main/java/nbc/ticketing/ticket911/application/seat/service/SeatService.java
@@ -2,22 +2,19 @@ package nbc.ticketing.ticket911.application.seat.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.LongStream;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
-import nbc.ticketing.ticket911.application.stage.service.StageService;
 import nbc.ticketing.ticket911.domain.seat.dto.request.CreateSeatRequestDto;
 import nbc.ticketing.ticket911.domain.seat.dto.request.UpdateSeatRequestDto;
 import nbc.ticketing.ticket911.domain.seat.dto.response.SeatResponseDto;
 import nbc.ticketing.ticket911.domain.seat.entity.Seat;
-import nbc.ticketing.ticket911.domain.seat.exception.SeatException;
-import nbc.ticketing.ticket911.domain.seat.exception.code.SeatExceptionCode;
 import nbc.ticketing.ticket911.domain.seat.repository.SeatRepository;
 import nbc.ticketing.ticket911.domain.seat.service.SeatDomainService;
 import nbc.ticketing.ticket911.domain.stage.entity.Stage;
+import nbc.ticketing.ticket911.domain.stage.service.StageDomainService;
 
 @Service
 @RequiredArgsConstructor
@@ -25,23 +22,14 @@ import nbc.ticketing.ticket911.domain.stage.entity.Stage;
 public class SeatService {
 
 	private final SeatRepository seatRepository;
-	private final StageService stageService;
 	private final SeatDomainService seatDomainService;
+	private final StageDomainService stageDomainService;
 
 	@Transactional
 	public List<SeatResponseDto> createSeat(Long stageId, CreateSeatRequestDto createSeatRequestDto) {
-		Stage stage = stageService.findStageByStageIdOrElseThrow(stageId);
+		Stage stage = stageDomainService.findStageByStageIdOrElseThrow(stageId);
 
-		List<Seat> seats = LongStream.range(0, createSeatRequestDto.getSeatCount())
-			.mapToObj(i -> {
-				Seat seat = Seat.builder()
-					.seatName(createSeatRequestDto.getSeatName())
-					.seatPrice(createSeatRequestDto.getSeatPrice())
-					.build();
-				seat.addStage(stage);
-				return seat;
-			})
-			.collect(Collectors.toList());
+		List<Seat> seats = seatDomainService.createSeats(createSeatRequestDto, stage);
 
 		seatRepository.saveAll(seats);
 		stage.updateTotalSeat(createSeatRequestDto.getSeatCount());
@@ -53,7 +41,7 @@ public class SeatService {
 	}
 
 	public List<SeatResponseDto> getSeats(Long stageId) {
-		Stage stage = stageService.findStageByStageIdOrElseThrow(stageId);
+		Stage stage = stageDomainService.findStageByStageIdOrElseThrow(stageId);
 		List<Seat> seats = stage.getSeats();
 
 		return seats.stream()
@@ -63,12 +51,10 @@ public class SeatService {
 
 	@Transactional
 	public SeatResponseDto updateSeat(Long stageId, Long seatId, UpdateSeatRequestDto updateSeatRequestDto) {
-		Stage stage = stageService.findStageByStageIdOrElseThrow(stageId);
-		Seat seat = getSeatBySeatIdOrElseThrow(seatId);
+		Stage stage = stageDomainService.findStageByStageIdOrElseThrow(stageId);
+		Seat seat = seatDomainService.getSeatBySeatIdOrElseThrow(seatId);
 
-		if (seat.getStage() != stage) {
-			throw new SeatException(SeatExceptionCode.SEAT_NOT_BELONG_TO_STAGE);
-		}
+		seatDomainService.verifySeat(seat, stage);
 
 		seat = seatDomainService.updateSeat(seat, updateSeatRequestDto);
 
@@ -78,20 +64,13 @@ public class SeatService {
 
 	@Transactional
 	public void deleteSeat(Long stageId, Long seatId) {
-		Stage stage = stageService.findStageByStageIdOrElseThrow(stageId);
-		Seat seat = getSeatBySeatIdOrElseThrow(seatId);
+		Stage stage = stageDomainService.findStageByStageIdOrElseThrow(stageId);
+		Seat seat = seatDomainService.getSeatBySeatIdOrElseThrow(seatId);
 
-		if (seat.getStage() != stage) {
-			throw new SeatException(SeatExceptionCode.SEAT_NOT_BELONG_TO_STAGE);
-		}
+		seatDomainService.verifySeat(seat, stage);
 
 		seatRepository.deleteById(seatId);
 
-	}
-
-	private Seat getSeatBySeatIdOrElseThrow(Long seatId) {
-		return seatRepository.findById(seatId)
-			.orElseThrow(() -> new SeatException(SeatExceptionCode.SEAT_NOT_FOUND));
 	}
 
 }
