+++ src/main/java/nbc/ticketing/ticket911/domain/stage/service/StageDomainService.java
@@ -2,11 +2,18 @@ package nbc.ticketing.ticket911.domain.stage.service;
 
 import org.springframework.stereotype.Component;
 
+import lombok.RequiredArgsConstructor;
 import nbc.ticketing.ticket911.domain.stage.entity.Stage;
+import nbc.ticketing.ticket911.domain.stage.exception.StageException;
+import nbc.ticketing.ticket911.domain.stage.exception.code.StageExceptionCode;
+import nbc.ticketing.ticket911.domain.stage.repository.StageRepository;
 import nbc.ticketing.ticket911.domain.stage.status.StageStatus;
 
 @Component
+@RequiredArgsConstructor
 public class StageDomainService {
+
+	private final StageRepository stageRepository;
 
 	public Stage createStage(String stageName) {
 		return Stage.builder()
@@ -14,5 +21,10 @@ public class StageDomainService {
 			.stageStatus(StageStatus.AVAILABLE)
 			.totalSeat(0L)
 			.build();
+	}
+
+	public Stage findStageByStageIdOrElseThrow(Long stageId) {
+		return stageRepository.findById(stageId)
+			.orElseThrow(() -> new StageException(StageExceptionCode.STAGE_NOT_FOUND));
 	}
 }
